requirement.txt
beautifulsoup4==4.13.4
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
gdown==5.2.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imgviz==1.7.6
kiwisolver==1.4.8
labelme==5.4.1
lazy_loader==0.4
matplotlib==3.10.1
mpmath==1.3.0
natsort==8.4.0
networkx==3.4.2
numpy==2.2.4
onnxruntime==1.15.0
packaging==24.2
pillow==11.2.1
protobuf==6.30.2
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
QtPy==2.4.3
requests==2.32.3
scikit-image==0.25.2
scipy==1.15.2
six==1.17.0
soupsieve==2.6
sympy==1.13.3
termcolor==3.0.1
tifffile==2025.3.30
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
